Add modules and webpack config
