+++ .gitignore
@@ -13,3 +13,7 @@ node_modules/
 
 # Папка с собранными файлами проекта
 build/
+
+# Bundle files
+public/bundle.js
+public/bundle.js.map
