+++ public/index.html
@@ -1923,6 +1923,7 @@ Here is a card with filled data</textarea
       </section>
     </main>
 
-    <script src="../src/main.js"></script>
+    <!-- <script src="../src/main.js"></script> -->
+    <script src="bundle.js"></script>
   </body>
 </html>
